Renamed & added compiled binaries for pi
